requirements-dev.txt
django-queryinspect
django-debug-toolbar
